Move Settings, SomethingWentWrong and TriggerChecksExecutionRequest
