+++ assets/js/pages/Settings/index.js
