+++ assets/js/trento.jsx
@@ -22,9 +22,9 @@ import HostDetailsPage, { HostSettingsPage } from '@pages/HostDetails';
 import DatabasesOverviewPage from '@pages/DatabasesOverview';
 import ChecksCatalogPage from '@pages/ChecksCatalog';
 import NotFound from '@pages/NotFound';
-import SomethingWentWrong from '@components/SomethingWentWrong';
+import SomethingWentWrong from '@pages/SomethingWentWrong';
 import SaptuneDetailsPage from '@pages/SaptuneDetails';
-import Settings from '@components/Settings';
+import Settings from '@pages/Settings';
 import Eula from '@pages/Eula';
 import Login from '@components/Login';
 import { me } from '@lib/auth';
